Add BLEU score calculation classes and losses module
